Remove "HTTP" from propagator names
These propagators are actually multi-purpose, not strictly for HTTP.
