+++ propagation/b3_propagator.go
@@ -35,7 +35,7 @@ const (
 	B3ParentSpanIDHeader = "X-B3-ParentSpanId"
 )
 
-// HTTPB3Propagator that facilitates core.SpanContext
+// B3Propagator that facilitates core.SpanContext
 // propagation using B3 Headers.
 // This propagator supports both version of B3 headers,
 //  1. Single Header :
@@ -49,13 +49,13 @@ const (
 //
 // If SingleHeader is set to true then X-B3 header is used to inject and extract. Otherwise,
 // separate headers are used to inject and extract.
-type HTTPB3Propagator struct {
+type B3Propagator struct {
 	SingleHeader bool
 }
 
-var _ apipropagation.TextFormatPropagator = HTTPB3Propagator{}
+var _ apipropagation.TextFormatPropagator = B3Propagator{}
 
-func (b3 HTTPB3Propagator) Inject(ctx context.Context, supplier apipropagation.Supplier) {
+func (b3 B3Propagator) Inject(ctx context.Context, supplier apipropagation.Supplier) {
 	sc := trace.CurrentSpan(ctx).SpanContext()
 	if sc.IsValid() {
 		if b3.SingleHeader {
@@ -79,21 +79,21 @@ func (b3 HTTPB3Propagator) Inject(ctx context.Context, supplier apipropagation.S
 }
 
 // Extract retrieves B3 Headers from the supplier
-func (b3 HTTPB3Propagator) Extract(ctx context.Context, supplier apipropagation.Supplier) (core.SpanContext, dctx.Map) {
+func (b3 B3Propagator) Extract(ctx context.Context, supplier apipropagation.Supplier) (core.SpanContext, dctx.Map) {
 	if b3.SingleHeader {
 		return b3.extractSingleHeader(supplier), dctx.NewEmptyMap()
 	}
 	return b3.extract(supplier), dctx.NewEmptyMap()
 }
 
-func (b3 HTTPB3Propagator) GetAllKeys() []string {
+func (b3 B3Propagator) GetAllKeys() []string {
 	if b3.SingleHeader {
 		return []string{B3SingleHeader}
 	}
 	return []string{B3TraceIDHeader, B3SpanIDHeader, B3SampledHeader}
 }
 
-func (b3 HTTPB3Propagator) extract(supplier apipropagation.Supplier) core.SpanContext {
+func (b3 B3Propagator) extract(supplier apipropagation.Supplier) core.SpanContext {
 	tid, err := core.TraceIDFromHex(supplier.Get(B3TraceIDHeader))
 	if err != nil {
 		return core.EmptySpanContext()
@@ -128,7 +128,7 @@ func (b3 HTTPB3Propagator) extract(supplier apipropagation.Supplier) core.SpanCo
 	return sc
 }
 
-func (b3 HTTPB3Propagator) extractSingleHeader(supplier apipropagation.Supplier) core.SpanContext {
+func (b3 B3Propagator) extractSingleHeader(supplier apipropagation.Supplier) core.SpanContext {
 	h := supplier.Get(B3SingleHeader)
 	if h == "" || h == "0" {
 		core.EmptySpanContext()
@@ -177,7 +177,7 @@ func (b3 HTTPB3Propagator) extractSingleHeader(supplier apipropagation.Supplier)
 }
 
 // extractSampledState parses the value of the X-B3-Sampled b3Header.
-func (b3 HTTPB3Propagator) extractSampledState(sampled string) (flag byte, ok bool) {
+func (b3 B3Propagator) extractSampledState(sampled string) (flag byte, ok bool) {
 	switch sampled {
 	case "", "0":
 		return 0, true
@@ -196,7 +196,7 @@ func (b3 HTTPB3Propagator) extractSampledState(sampled string) (flag byte, ok bo
 }
 
 // extracDebugFlag parses the value of the X-B3-Sampled b3Header.
-func (b3 HTTPB3Propagator) extracDebugFlag(debug string) (flag byte, ok bool) {
+func (b3 B3Propagator) extracDebugFlag(debug string) (flag byte, ok bool) {
 	switch debug {
 	case "", "0":
 		return 0, true
