+++ propagation/trace_context_propagator_benchmark_test.go
@@ -11,7 +11,7 @@ import (
 )
 
 func BenchmarkInject(b *testing.B) {
-	var t HTTPTraceContextPropagator
+	var t TraceContextPropagator
 
 	injectSubBenchmarks(b, func(ctx context.Context, b *testing.B) {
 		req, _ := http.NewRequest("GET", "http://example.com", nil)
@@ -52,7 +52,7 @@ func injectSubBenchmarks(b *testing.B, fn func(context.Context, *testing.B)) {
 
 func BenchmarkExtract(b *testing.B) {
 	extractSubBenchmarks(b, func(b *testing.B, req *http.Request) {
-		var propagator HTTPTraceContextPropagator
+		var propagator TraceContextPropagator
 		ctx := context.Background()
 		b.ResetTimer()
 		for i := 0; i < b.N; i++ {
